Separate tests into each categories
