+++ package.json
@@ -415,7 +415,7 @@
 		"pretest": "npm run compile && npm run lint",
 		"lint": "eslint src --ext ts",
 		"test": "node ./out/test/runTest.js",
-		"ui-test": "extest setup-and-run -i out/test/spec/*test.js"
+		"ui-test": "extest setup-and-run -i out/test/*.test.js"
 	},
 	"extensionDependencies": [
 		"ms-vscode.cpptools",
